test Reservation #202
some temp implementations were added

+++ src/main/java/il/org/parking/Price.java
@@ -10,13 +10,13 @@ public interface Price {
   /**
    * This method is used to inform the user of the price of 
    * some object.
-   * @return Integer which represents the price of the object
+   * @return Double which represents the price of the object
    */
-  Integer getCost();
+  Double getCost();
   
   /**
    * This method is used to update the price data to a new value
-   * @param Integer which represents the new price.
+   * @param Double which represents the new price.
    */
-  void update(Integer newPrice);
+  void update(Double newPrice);
 }
